Start of integrating a mode setup into virtqueue.
This commit starts the integration of setting-up a queue in different
modes upon startup. This will allow the user of the queue how he wants
to be "notified" by the queue.

Two usecases currently are in mind.
1. Poll Mode: The user will actively poll the queue via a poll function.
Unless the queue is not polled, no descriptors are processed by the
queue.

2. Busy Mode: The queue starts a second thread/task which permanently
polls if any descriptors have been processed and will set the respective
TransferToken.state == TransferState::Finished. This mode works best if
the TransferToken.await_queue is set, as the user of the queue can
simply look at the VecDequeue<Transfer> if any Transfers are finished.

In both cases, the queue will push finshed Transfers into the
TransferToken.await_queue if it is available.